compile fix

+++ JUCE/modules/juce_gui_basics/native/juce_Windowing_mac.mm
@@ -431,7 +431,7 @@ static Displays::Display getDisplayFromScreen (NSScreen* s, CGFloat& mainScreenB
 {
     Displays::Display d;
 
-    d.isMain = (approximatelyEqual (mainScreenBottom, 0.0));
+    d.isMain = (approximatelyEqual ((double) mainScreenBottom, 0.0));
 
     if (d.isMain)
         mainScreenBottom = [s frame].size.height;
